Nästan klar med endpoint api-status

+++ types/api_status.go
@@ -8,10 +8,10 @@ type ApiStatus struct {
 }
 
 type Uptime struct {
-	Years   uint8 `json:"years"`
-	Months  uint8 `json:"months"`
-	Days    uint8 `json:"days"`
-	Hours   uint8 `json:"hours"`
-	Minutes uint8 `json:"minutes"`
-	Seconds uint8 `json:"seconds"`
+	Years   uint16 `json:"years"`
+	Months  uint8  `json:"months"`
+	Days    uint16 `json:"days"`
+	Hours   uint8  `json:"hours"`
+	Minutes uint8  `json:"minutes"`
+	Seconds uint8  `json:"seconds"`
 }
